requirements.txt
altgraph==0.17.4
click==8.1.7
colorama==0.4.6
MouseInfo==0.1.3
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pefile==2023.2.7
pillow==10.4.0
platformdirs==4.2.2
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.7
PyPDF2==3.0.1
pyperclip==1.9.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.2
PyQt6_sip==13.8.0
PyRect==0.2.0
PyScreeze==0.1.30
pytweening==1.2.0
pywin32-ctypes==0.2.2
setuptools==72.1.0
six==1.16.0
